Refactoring subscriber
