+++ litmus-portal/cluster-agents/subscriber/pkg/cluster/logs/pod.go
@@ -4,7 +4,7 @@ import (
 	"bytes"
 	"io"
 
-	"github.com/litmuschaos/litmus/litmus-portal/backend/subscriber/pkg/k8s"
+	"github.com/litmuschaos/litmus/litmus-portal/cluster-agents/subscriber/pkg/k8s"
 
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/client-go/kubernetes"
