+++ litmus-portal/cluster-agents/subscriber/pkg/gql/subscriptions.go
@@ -4,8 +4,8 @@ import (
 	"encoding/json"
 
 	"github.com/gorilla/websocket"
-	"github.com/litmuschaos/litmus/litmus-portal/backend/subscriber/pkg/cluster/operations"
-	"github.com/litmuschaos/litmus/litmus-portal/backend/subscriber/pkg/types"
+	"github.com/litmuschaos/litmus/litmus-portal/cluster-agents/subscriber/pkg/cluster/operations"
+	"github.com/litmuschaos/litmus/litmus-portal/cluster-agents/subscriber/pkg/types"
 	"github.com/sirupsen/logrus"
 
 	"log"
