+++ litmus-portal/cluster-agents/subscriber/pkg/types/event.go
